feat: remove microfrontend poc
