+++ src/index.tsx
@@ -1,1 +1,12 @@
-import("./bootstrap");
+import React from "react";
+import App from "./App";
+import { createRoot } from "react-dom/client";
+
+const container = document.getElementById("app-root")!;
+const root = createRoot(container);
+
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
