+++ webpack.config.js
@@ -1,8 +1,5 @@
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
-const ModuleFederationPlugin = require("webpack/lib/container/ModuleFederationPlugin");
-
-const deps = require("./package.json").dependencies;
 
 module.exports = {
   mode: "development",
@@ -42,25 +39,6 @@ module.exports = {
   },
   devtool: "source-map",
   plugins: [
-    new ModuleFederationPlugin({
-      name: "shop_ui",
-      filename: "remoteEntry.js",
-      remotes: {
-        ui: "ui@http://localhost:3000/remoteEntry.js",
-      },
-      exposes: {},
-      shared: {
-        ...deps,
-        react: {
-          singleton: true,
-          requiredVersion: deps.react,
-        },
-        "react-dom": {
-          singleton: true,
-          requiredVersion: deps["react-dom"],
-        },
-      },
-    }),
     new HtmlWebpackPlugin({
       template: "index.html",
     }),
